--- a/include/SFML/Graphics/Transform.inl
+++ b/include/SFML/Graphics/Transform.inl
@@ -41,12 +41,11 @@ constexpr Transform::Transform() = default;
 ////////////////////////////////////////////////////////////
 // clang-format off
 constexpr Transform::Transform(float a00, float a01, float a02,
-                               float a10, float a11, float a12,
-                               float a20, float a21, float a22)
-    : m_matrix{a00, a10, 0.f, a20,
-               a01, a11, 0.f, a21,
+                               float a10, float a11, float a12)
+    : m_matrix{a00, a10, 0.f, 0.f,
+               a01, a11, 0.f, 0.f,
                0.f, 0.f, 1.f, 0.f,
-               a02, a12, 0.f, a22}
+               a02, a12, 0.f, 1.f}
 {
 }
 // clang-format on
@@ -64,9 +63,7 @@ constexpr Transform Transform::getInverse() const
 {
     // clang-format off
     // Compute the determinant
-    const float det = m_matrix[0] * (m_matrix[15] * m_matrix[5] - m_matrix[7] * m_matrix[13]) -
-                      m_matrix[1] * (m_matrix[15] * m_matrix[4] - m_matrix[7] * m_matrix[12]) +
-                      m_matrix[3] * (m_matrix[13] * m_matrix[4] - m_matrix[5] * m_matrix[12]);
+    const float det = m_matrix[0] * m_matrix[5] - m_matrix[1] * m_matrix[4];
     // clang-format on
 
     // Compute the inverse if the determinant is not zero
@@ -74,15 +71,12 @@ constexpr Transform Transform::getInverse() const
     if (det != 0.f)
     {
         // clang-format off
-        return {(m_matrix[15] * m_matrix[5] - m_matrix[7] * m_matrix[13]) / det,
-               -(m_matrix[15] * m_matrix[4] - m_matrix[7] * m_matrix[12]) / det,
-                (m_matrix[13] * m_matrix[4] - m_matrix[5] * m_matrix[12]) / det,
-               -(m_matrix[15] * m_matrix[1] - m_matrix[3] * m_matrix[13]) / det,
-                (m_matrix[15] * m_matrix[0] - m_matrix[3] * m_matrix[12]) / det,
-               -(m_matrix[13] * m_matrix[0] - m_matrix[1] * m_matrix[12]) / det,
-                (m_matrix[7]  * m_matrix[1] - m_matrix[3] * m_matrix[5])  / det,
-               -(m_matrix[7]  * m_matrix[0] - m_matrix[3] * m_matrix[4])  / det,
-                (m_matrix[5]  * m_matrix[0] - m_matrix[1] * m_matrix[4])  / det};
+		return {(               m_matrix[5]                             ) / det,
+               -(               m_matrix[4]                             ) / det,
+				(m_matrix[13] * m_matrix[4] - m_matrix[5] * m_matrix[12]) / det,
+               -(               m_matrix[1]                             ) / det,
+                (               m_matrix[0]                             ) / det,
+               -(m_matrix[13] * m_matrix[0] - m_matrix[1] * m_matrix[12]) / det};
         // clang-format on
     }
     else
@@ -137,15 +131,12 @@ constexpr Transform& Transform::combine(const Transform& transform)
     const auto& b = transform.m_matrix;
 
     // clang-format off
-    *this = Transform(a[0] * b[0]  + a[4] * b[1]  + a[12] * b[3],
-                      a[0] * b[4]  + a[4] * b[5]  + a[12] * b[7],
-                      a[0] * b[12] + a[4] * b[13] + a[12] * b[15],
-                      a[1] * b[0]  + a[5] * b[1]  + a[13] * b[3],
-                      a[1] * b[4]  + a[5] * b[5]  + a[13] * b[7],
-                      a[1] * b[12] + a[5] * b[13] + a[13] * b[15],
-                      a[3] * b[0]  + a[7] * b[1]  + a[15] * b[3],
-                      a[3] * b[4]  + a[7] * b[5]  + a[15] * b[7],
-                      a[3] * b[12] + a[7] * b[13] + a[15] * b[15]);
+    *this = Transform(a[0] * b[0]  + a[4] * b[1]         ,
++                      a[0] * b[4]  + a[4] * b[5]         ,
++                      a[0] * b[12] + a[4] * b[13] + a[12],
++                      a[1] * b[0]  + a[5] * b[1]         ,
++                      a[1] * b[4]  + a[5] * b[5]         ,
++                      a[1] * b[12] + a[5] * b[13] + a[13]);
     // clang-format on
 
     return *this;
@@ -157,8 +148,7 @@ constexpr Transform& Transform::translate(const Vector2f& offset)
 {
     // clang-format off
     const Transform translation(1, 0, offset.x,
-                                0, 1, offset.y,
-                                0, 0, 1);
+                                0, 1, offset.y);
     // clang-format on
 
     return combine(translation);
@@ -170,8 +160,7 @@ constexpr Transform& Transform::scale(const Vector2f& factors)
 {
     // clang-format off
     const Transform scaling(factors.x, 0,         0,
-                            0,         factors.y, 0,
-                            0,         0,         1);
+                            0,         factors.y, 0);
     // clang-format on
 
     return combine(scaling);
@@ -183,8 +172,7 @@ constexpr Transform& Transform::scale(const Vector2f& factors, const Vector2f& c
 {
     // clang-format off
     const Transform scaling(factors.x, 0,         center.x * (1 - factors.x),
-                            0,         factors.y, center.y * (1 - factors.y),
-                            0,         0,         1);
+                            0,         factors.y, center.y * (1 - factors.y));
     // clang-format on
 
     return combine(scaling);
@@ -219,9 +207,9 @@ constexpr bool operator==(const Transform& left, const Transform& right)
     const float* b = right.getMatrix();
 
     // clang-format off
-    return ((a[0]  == b[0])  && (a[1]  == b[1])  && (a[3]  == b[3]) &&
-            (a[4]  == b[4])  && (a[5]  == b[5])  && (a[7]  == b[7]) &&
-            (a[12] == b[12]) && (a[13] == b[13]) && (a[15] == b[15]));
+    return ((a[0]  == b[0])  && (a[1]  == b[1])
+	     && (a[4]  == b[4])  && (a[5]  == b[5])
+		 && (a[12] == b[12]) && (a[13] == b[13]));
     // clang-format on
 }