--- a/test/Graphics/RenderWindow.test.cpp
+++ b/test/Graphics/RenderWindow.test.cpp
@@ -43,7 +43,7 @@ TEST_CASE("[Graphics] sf::RenderWindow", runDisplayTests())
             CHECK(window.getView().getSize() == sf::Vector2f(256, 256));
             CHECK(window.getView().getRotation() == sf::Angle::Zero);
             CHECK(window.getView().getViewport() == sf::FloatRect({0, 0}, {1, 1}));
-            CHECK(window.getView().getTransform() == sf::Transform(0.0078125f, 0, -1, 0, -0.0078125f, 1, 0, 0, 1));
+            CHECK(window.getView().getTransform() == sf::Transform(0.0078125f, 0, -1, 0, -0.0078125f, 1));
         }
 
         SECTION("State and settings")
@@ -62,7 +62,7 @@ TEST_CASE("[Graphics] sf::RenderWindow", runDisplayTests())
             CHECK(window.getView().getRotation() == sf::Angle::Zero);
             CHECK(window.getView().getViewport() == sf::FloatRect({0, 0}, {1, 1}));
             CHECK(window.getView().getTransform() ==
-                  Approx(sf::Transform(0.00833333f, 0, -1, 0, -0.00666667f, 1, 0, 0, 1)));
+                  Approx(sf::Transform(0.00833333f, 0, -1, 0, -0.00666667f, 1)));
         }
     }