--- a/test/Graphics/Transform.test.cpp
+++ b/test/Graphics/Transform.test.cpp
@@ -8,7 +8,6 @@
 #include <sstream>
 #include <type_traits>
 #include <vector>
-#include <iostream>
 
 #include <cassert>