<!--
Thanks a lot for making a contribution to SFML! 🙂

Before you create the pull request, we ask you to check the follow boxes. (For small changes not everything needs to ticked, but the more the better!)

-   [x] Has this change been discussed on [the forum](https://en.sfml-dev.org/forums/index.php#c3) or in an issue before? (discord)
-   [x] Does the code follow the SFML [Code Style Guide](https://www.sfml-dev.org/style.php)?
-   [x] Have you provided some example/test code for your changes?
-   [ ] If you have additional steps which need to be performed list them as tasks!
-->

## Description

Earlier I noted that all the constructors of `sf::Transform` in SFML follow the property of the bottom row being 0, 0, 1.  
All available operations on any `Transform`s with this property, results in the property also holding true for the resulting `Transform`. This includes the dot product `Transform::combine()` and `Transform.getInverse()`.  
The only way to break this property, is if the API user manually uses the `Transform::Transform(...)` constructor to create a Transform with a different bottom row. I could not find any useseful use cases for this.

This property can be exploited to significantly optimize the math of the operators of `Transform`. Especially in the heavily used `Transform::combine()` but also notably in `Transform::getInverse()`.

I did this because I could on a branch and have been using it in my own game for some time now. I understand that this is the lowest priority of low priority PR's, but since I already made the change I at least wanted to bring up the discussion rather than letting it go to waste.

## Tasks

-   [ ] Tested on Linux
-   [X] Tested on Windows
-   [ ] Tested on macOS
-   [ ] Tested on iOS
-   [ ] Tested on Android

## How to test this PR?

Open for suggestions.