@ChrisThrasher 2024-05-20
Are you able to write a benchmark to demonstrate the extent to which this improves performance?