@eXpl0it3r 2024-05-21
Not exactly a fan of this, as we're moving away from the mathematical definition in the name of some (unproven) "optimization".

Until now `sf::Transform` was really just a 4x4 transformation matrix. Now it's just "something". Any interoperability where a normal transformation matrix is expected, would need to readd the bottom row. I'm not even sure if the conversion to `sf::Glsl::Mat4` is correct now.

> The only way to break this property, is if the API user manually uses the Transform::Transform(...) constructor to create a Transform with a different bottom row. I could not find any useseful use cases for this.

For a normal transform changing the value might not matter, but I don't see why we should remove the mathematical possibilities from users.

If I understood this correctly, if you change the last row, you're no longer transforming within the 2D plane (R^2), but would map it into another space.
- https://math.stackexchange.com/questions/335932/what-is-the-sense-of-bottom-row-of-affine-transform-matrix
- https://stackoverflow.com/questions/32565827/whats-the-purpose-of-magic-4-of-last-row-in-matrix-4x4-for-3d-graphics