@vittorioromeo 2024-05-21
I like the simplicity the optimization behind this PR, but the only way I would be comfortable accepting this is if instead of `sf::Transform` you introduced a brand new "transform type" that is convertible to `sf::Transform` that is only used internally by SFML.

`sf::Transform` is a general-purpose transformation matrix that is part of the public API, we cannot predict if users will always have a bottom row of `[0, 0, 1]`.

I think it's worthwhile investigating if all internal uses of `sf::Transform` by SFML can be changed to some `sf::priv::Transform2D` instead that has an hardcoded bottom row of `[0, 0, 1]`.