@L0laapk3 2024-05-21
> Not exactly a fan of this, as we're moving away from the mathematical definition in the name of some (unproven) "optimization".

I believe there is some misunderstanding here. I did not "remove" the bottom row from the calculations. I made it have fixed the values to 0, 0, 1 rather than them being configurable.

1. It is a mathematically proveable fact that when using only square matrices with a bottom row of `0, .., 0, 1`, for all the `sf::Transform` operations, the result is always another matrix with a bottom row of `0, ..., 0, 1`.