@L0laapk3 2024-05-21
> Not exactly a fan of this, as we're moving away from the mathematical definition in the name of some (unproven) "optimization".

I believe there is some misunderstanding here. I did not "remove" the bottom row from the calculations. I made it have fixed the values to `0 0 1` rather than them being configurable.

1. It is a mathematically proveable fact that when using only square matrices with a bottom row of `0 .. 0 1`, for all the `sf::Transform` operations, the result is always another matrix with a bottom row of `0 .. 0 1`. I encourage you to fill in these values yourself for the bottom row and observe that in all cases the bottom row becomes `0 0 1` again and you will end up with the exact same functions as present in the PR.
2. It is easily checkable that indeed all code paths only construct `sf::Transform`s with a bottom row of `0 0 1`. The one notable exception being the freeform `sf::Transform(a00, a01, .., a21, a22)` constructor which I propose to remove this flexibility from (See below on what functionality is removed).

> Until now sf::Transform was really just a 4x4 transformation matrix.

This is not true, the third row and column in the 4x4 underlying memory object are already hardcoded to identity values, and these values were already optimized out of all the operators. As far as all mathematical operations are concerned, `sf::Transform` behaves as a 3x3 matrix. This PR does not change that into ""something"". It still behaves mathematically as a 3x3 matrix, only with the constraint that the bottom row can not be changed to anything other than `0 0 1`.

> If I understood this correctly, if you change the last row, you're no longer transforming within the 2D plane (R^2), but would map it into another space.

This is incorrect since I did not "change" the bottom row, at least in a mathematical sense. It just restricts the values to be `0 0 1`. `sf::Transform`. In the mathematical sense, `sf::Transform` is now a 2D affine transformation matrix able to do all affine transformations.

> For a normal transform changing the value might not matter, but I don't see why we should remove the mathematical possibilities from users.

**What functionality is actually lost by restricting the bottom row to `0 0 1`?**

As mentioned, all affine transformations are still supported (This includes all scaling, translation, rotation). So functionality is actually lost by not supporting other linear transformations?

One non-affine use case is projections. In a 3D environment, 4x4 non-affine transformations can be used to project a 3D space onto a 2D plane (Read: a non-isometric 2D camera of a 3D world). Since SFML already removes a dimension, this would mean with this change it is no longer possible to project a 2D space onto a 1D (line) camera in a non-isometric way.

I did also find this demo, where someone made some artistic warping effects: https://github.com/IRCSS/Non-Affine-Transformation-Matrices/blob/master/Assets/Shaders/Resources/nonAffineBending.shader. As far as I can tell, this is neither an affine nor a linear transformation, he uses a different (parametrized by theta) affine transformation for each vertex in his model, and this is not less possible after this PR compared to before this PR.