@eXpl0it3r 2024-05-21
The first step of any optimization discussion is to do actual benchmarking. If they don't yield a significant difference, there's no need in discussing anything further, even if the calculation is "technically" simplified.

> the only way I would be comfortable accepting this is if instead of `sf::Transform` you introduced a brand new "transform type" that is convertible to `sf::Transform` that is only used internally by SFML.

I'm generally not a fan of introducing internal constructs for optimizations like this. You're trading complexity for (potential) performance gains. I much rather keep the complexity low and defer people to highly optimized math libraries if they require performance.