@binary1248 2024-05-24
So these are the results I got when building with VS2022 ClangCL x64 and specifying flags `/arch:AVX512 /Gv /fp:fast` to give the optimizer as much freedom as it needs to get work done. Run on a Ryzen 9 7950X.
### Old
```
-------------------------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------
BM_transform_combine/min_time:4.000          5.30 ns         5.28 ns   1054117647 items_per_second=189.504M/s
BM_transform_getInverse/min_time:4.000       1.30 ns         1.30 ns   4318072289 items_per_second=767.657M/s
BM_transform_translate/min_time:4.000        4.66 ns         4.66 ns   1194666667 items_per_second=214.772M/s
BM_transform_rotate/min_time:4.000           8.56 ns         8.57 ns    676226415 items_per_second=116.654M/s
BM_transform_scale/min_time:4.000            4.99 ns         4.98 ns   1120000000 items_per_second=200.784M/s
```
### New
```
-------------------------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------
BM_transform_combine/min_time:4.000          5.25 ns         5.25 ns   1054117647 items_per_second=190.575M/s
BM_transform_getInverse/min_time:4.000       1.12 ns         1.12 ns   4977777778 items_per_second=892.375M/s
BM_transform_translate/min_time:4.000        4.69 ns         4.70 ns   1194666667 items_per_second=212.977M/s
BM_transform_rotate/min_time:4.000           7.78 ns         7.78 ns    702745098 items_per_second=128.502M/s
BM_transform_scale/min_time:4.000            4.37 ns         4.38 ns   1280000000 items_per_second=228.189M/s
```

Just a disclaimer up front, normally people will know me as the guy who has optimized performance in many places in the past. One rule that I never broke however, is that any optimization I perform must never have an effect on the public API of the library.

When writing a library there are multiple goals that have to be kept in mind. Obviously, a library that has horrible performance is not going to be one that many people will be willing to use. But a library that people have a hard time learning/adopting will also be a huge turn off. Keeping all the design and implementation goals in balance is stuff library maintainers have to deal with all the time when making changes.

In this case, your benchmarks have proven that at an instruction level, yes, forcing the CPU to run less instructions does have an impact on the performance of these functions when observing them under a microscope. However, we must not forget about considering the global context: where are these functions used, and is this really a bottleneck in real-world non-synthetic benchmarks?

The main reason why `sf::Transform` exists is because sfml-graphics was built on top of legacy OpenGL. In legacy OpenGL the only way users could _transform_ vertices without manipulating vertex data was by using the provided matrix stack. Rotations, translations and scaling would all be performed by the CPU and the final matrix fed to OpenGL. This was the model from the beginnings of OpenGL in the 1990s, when GPUs weren't really that specialized in doing matrix maths and not nearly as massively parallel as they are today. All of this maths we are trying to optimize here really belongs on the GPU where it was meant to be run. The only reason we are still where we are right now is because we maintainers just haven't gotten around to modernizing the graphics module just yet.

Coming back to how and where these functions are used, the only place where `::combine` is currently used in the library is by the other `Transform` functions. When the user decides to transform a `Transformable` they end up using one of either `::rotate`, `::translate` or `::scale` never actually calling `::combine` directly. As you said all of theses operations are affine transformations, so indeed we could optimize the maths behind them if we know an entire row is always fixed. Now comes the question: if it is so obvious to us that the maths can be simplified, why hasn't the compiler performed the optimizations for us and thrown out some of the unnecessary instructions?

From arithmetic we learned in school we know that `0 * x = 0` and `1 * x = x`. It just so happens the compiler also knows these rules... _for integer arithmetic_. When we dive into IEEE 754 floating point arithmetic we get `0 * x = <it depends>` and `1 * x = <it depends>`. These are not rules that are helpful to the optimizer. If anyone wants to read up more on this, you can do so [here](https://en.wikipedia.org/wiki/IEEE_754). To cut a long story short, IEEE 754 requires special handling of "special values" that breaks the rules we learned in school. We can see this in action using [compiler explorer](https://gcc.godbolt.org/z/6baPG9Y3v). Only when telling the compiler that it can perform "fast floating point operations" can it treat floats using the same arithmetic rules as it does for integers.

What does this all mean for us? Well, if we were to be really really pedantic, the changes you made would actually lead to different results if we fed it IEEE 754 special values. Does it really matter for everyday use? Probably not, but that is not a decision for the compiler to make, hence us having to specify explicitly that we don't care using the appropriate compiler flags.

If we can assume that we only have to work with "normal" values, we can still optimize the performance when performing the affine transformations, all we need to do is... be the optimizer ourselves. What you have done by hand is what the optimizer would have done if it was allowed to but could not because we passively forbid it. If we performed the inlining and eliding of mathematical terms ourselves we would end up with the performance that your code demonstrates without having to touch `::combine` and thus leaving the public API as it is. I think this would be a change that would encounter less resistance than the current proposal.

Additionally, the benchmark results I ran were built using ClangCL as noted above. Why didn't I build them using the Microsoft compiler? Well, what I saw confirmed the notorious reputation that the Microsoft compiler is just _really really bad_ at inlining code where it would make complete sense. Being able to inline code is also a prerequisite for a number of other optimizations, which also wouldn't be able to be performed by the Microsoft compiler, optimizations like auto-vectorization using SIMD instructions like [FMA](https://en.wikipedia.org/wiki/FMA_instruction_set). Basically, the `Transform` code as it is written now exploits the "Achilles Heel" of the Microsoft compiler if you can call it that, by making a call to `::combine` instead of performing the operations inline within the function body. Because it is a complete function call, the function body will have to have a prologue and epilogue that also costs CPU instructions, the function arguments will have to be passed using indirect memory accesses if they can't all fit in registers, etc. etc. This is the reason why I tried building with `/Gv` forcing the compiler to use the `__vectorcall` convention making the call overhead less of a problem. When building with ClangCL, this was not a problem because the `::combine` code was always inlined.

Taking it further, I told ClangCL that it could use up to AVX512 instructions to optimize the maths that was being performed. This actually worked out pretty well. When inspecting the instructions that were generated, you could see it was using a bunch of FMA instructions to perform the multiplications and additions. However if you look closer, and use tools like AMD uProf to instrument the code with the help of CPU counters, the next problem arises.

Because we aren't performing a "real" 4x4 matrix multiply, instead choosing to skip certain rows/columns we are only performing 3x32bit = 96bit operations. Since the SSE instruction set was introduced, SIMD registers have always been at least 128 bits wide. This means that you would perform SIMD operations optimally in groups of 4 floats. We can't tell the processor to ignore a value because that's... not how it works. Now you might be asking, if we only work on groups of 3 floats, how does ClangCL still manage to make use of 128bit instructions? Well, it just has to make sure that the memory/register that wasn't meant to be touched is appropriately restored. Compilers know the cost of each instruction, and if they deem that it is still a better option to run a 128bit instruction + restore instead of 4 32bit instructions it will generate the appropriate code. If you think of it, this is how [SMR hard drives](https://en.wikipedia.org/wiki/Shingled_magnetic_recording) also write their data, and everyone knows their performance is worse than that of [CMR hard drives](https://en.wikipedia.org/wiki/Perpendicular_recording).

What I am trying to say is that, sometimes, we shouldn't try to be too clever. Trying to help the compiler save work only leads to it being able to perform less optimizations on its own. Sometimes, doing more work actually takes _less time_ because of the "weird" ways in which modern processors work. In the end, without diving into the generated instructions using e.g. compiler explorer and instrumenting using actual CPU counters, any performance optimizations are at best "hand wavy" and at worst can actually lead to worse performance when enabling all the instruction set optimizations that are available to the compiler. Without diving into the whole "optimizing debug performance" discussion that is ongoing, I myself am not a big fan of optimizing... well... non-optimized code. The benchmark data I showed above shows that the gains become marginal when all compiler optimizations are actually enabled.

As a general rule of thumb, before performing any kind of optimizations, I like to ask myself if the problem I am trying to solve has been solved by countless others before me. In this case, multiplying 4x4 matrices on CPU is probably one of those solved problems, for which the CPU manufacturers have even helped us over the years. We just have to make use of the new tools they have provided us to get the work done. If developers know they are going to be performing a lot of linear algebra, I don't think SFML is the library of choice either. There are a multitude of better linear algebra libraries out there that even contain hand-written assembly that exploits all of the aforementioned optimizations. Here at SFML, performance is really good enough for us as long as it doesn't become a bottleneck in real-world scenarios.

To sum up, like I said, if you were to inline your optimizations into `::rotate`, `::scale` and `::translate` and manually simplify the computations there, we would be able to benefit from your performance gains without having to change the public API. This is a change that @eXpl0it3r and I would take less issue with than the current proposal that touches the public API.