@binary1248 2024-05-24
Like I said, I ran the benchmark on Windows directly (not WSL), built with VS2022 using ClangCL because the default compiler produced even slower results. I don't know why you assume that the Ryzen 9 7950X and Intel 13900KF should produce the exact same results when built with different compilers, especially in synthetic tests. It is known that especially in synthetic single-threaded tests where everything fits in core-local L1D cache, Intel always came out in front because of their higher L1D bandwidth. AMD favours having larger shared L3 because in their philosophy it benefits real-world workloads more since they are becoming more multi-threaded these days. What makes matters worse is that, like I explained in my previous comment, we have a high amount of unaligned accesses here due to the way the code is written, another benefit for the Intel cache architecture. There have been enough articles on exactly this topic, but doing a quick search I found [this good one](https://chipsandcheese.com/2022/11/08/amds-zen-4-part-2-memory-subsystem-and-conclusion/).

Just for sake of discussion, I ran the benchmarks on a Ryzen 7 3600 (I don't have access to any Intel processors) running Debian and built with GCC and `-mavx2 -ffast-math` and these are the results.
### Old
```
-------------------------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------
BM_transform_combine/min_time:4.000          4.70 ns         4.70 ns   1190730302 items_per_second=212.782M/s
BM_transform_getInverse/min_time:4.000       4.84 ns         4.84 ns   1164881903 items_per_second=206.803M/s
BM_transform_translate/min_time:4.000        3.42 ns         3.42 ns   1628767426 items_per_second=292.309M/s
BM_transform_rotate/min_time:4.000           14.1 ns         14.1 ns    393183950 items_per_second=70.8252M/s
BM_transform_scale/min_time:4.000            3.18 ns         3.18 ns   1758419394 items_per_second=314.159M/s
```
### New
```
-------------------------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------
BM_transform_combine/min_time:4.000          2.74 ns         2.74 ns   2044065744 items_per_second=365.195M/s
BM_transform_getInverse/min_time:4.000       2.83 ns         2.83 ns   1974025570 items_per_second=353.804M/s
BM_transform_translate/min_time:4.000        2.69 ns         2.69 ns   2084774380 items_per_second=372.146M/s
BM_transform_rotate/min_time:4.000           12.4 ns         12.4 ns    452448636 items_per_second=80.8359M/s
BM_transform_scale/min_time:4.000            2.45 ns         2.45 ns   2279562321 items_per_second=407.835M/s
```
Because the difference is larger when run on a processor with higher memory latency (Zen 2 vs Zen 4) I can only assume that it means that the `Transform` code is pretty heavy on L1 accesses, especially misaligned ones as AMD uProf reported. Obviously removing L1 accesses as your code does will have a larger impact the worse the latency/bandwidth will be.

It is also a known fact that for CPU intensive workloads, Linux tends to win over Windows for various reasons.

When optimizing code, I wouldn't get too hung up on the absolute values that the benchmarks provide. As we have seen, it is highly dependent on compiler, operating system and CPU. What still has value however is a side-to-side performance comparison when the benchmarks are run in an identical environment. Some environments may have other bottlenecks which might skew the results in one direction or the other, but I would not label them "suspicious" just because they don't fit my expectations. Not everyone is going to run their SFML application on a 13900KF in WSL compiled with Clang (especially considering that SFML has known issues running in WSL anyway). Cherry picking results due to confirmation bias isn't really healthy for this kind of discussion considering many SFML users will be running their applications on Windows, probably built with the standard Microsoft compiler, running on perhaps even older AMD hardware.

If there is anyone out there with a 7950X running Windows, I have provided all the information necessary to be able to hopefully reproduce my results.